backend-py/requirements.txt
# Generated by: pip freeze (from requirements.in)
# To update: pip install -r requirements.in && pip freeze > requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
et_xmlfile==2.0.0
fastapi==0.133.1
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.7.3
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rsa==4.9.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
